# src/main/java/com/example/bankcards/dto/CardResponseDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.CardStatus;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDate;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class CardResponseDto {
    private Long id;
    private String maskedNumber;
    private String owner;
    private LocalDate expirationDate;
    private CardStatus status;
    private BigDecimal balance;
}